[TAN-3197] Export and reuse the UserLink component
