+++ front/app/containers/MainHeader/Components/NotificationMenu/components/CommentMarkedAsSpamNotification/index.tsx
@@ -5,14 +5,12 @@ import { ICommentMarkedAsSpamNotificationData } from 'api/notifications/types';
 import T from 'components/T';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import Link from 'utils/cl-router/Link';
-import { isNilOrError, stopPropagation } from 'utils/helperUtils';
 
 // data
 
 import messages from '../../messages';
-import { DeletedUser } from '../Notification';
 import NotificationWrapper from '../NotificationWrapper';
+import UserLink from '../UserLink';
 
 interface Props {
   notification: ICommentMarkedAsSpamNotificationData;
@@ -20,10 +18,6 @@ interface Props {
 
 const CommentMarkedAsSpamNotification = memo<Props>((props) => {
   const { notification } = props;
-
-  const deletedUser =
-    isNilOrError(notification.attributes.initiating_user_first_name) ||
-    isNilOrError(notification.attributes.initiating_user_slug);
 
   return (
     <NotificationWrapper
@@ -35,17 +29,11 @@ const CommentMarkedAsSpamNotification = memo<Props>((props) => {
       <FormattedMessage
         {...messages.userReportedCommentAsSpam}
         values={{
-          name: deletedUser ? (
-            <DeletedUser>
-              <FormattedMessage {...messages.deletedUser} />
-            </DeletedUser>
-          ) : (
-            <Link
-              to={`/profile/${notification.attributes.initiating_user_slug}`}
-              onClick={stopPropagation}
-            >
-              {notification.attributes.initiating_user_first_name}
-            </Link>
+          name: (
+            <UserLink
+              userName={notification.attributes.initiating_user_first_name}
+              userSlug={notification.attributes.initiating_user_slug}
+            />
           ),
           postTitle: <T value={notification.attributes.post_title_multiloc} />,
         }}
