+++ front/app/containers/MainHeader/Components/NotificationMenu/components/CommentOnIdeaYouFollowNotification/index.tsx
@@ -9,13 +9,12 @@ import { getInputTerm } from 'api/phases/utils';
 import useProjectById from 'api/projects/useProjectById';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import Link from 'utils/cl-router/Link';
-import { isNilOrError, stopPropagation } from 'utils/helperUtils';
+import { isNilOrError } from 'utils/helperUtils';
 import { getInputTermMessage } from 'utils/i18n';
 
 import messages from '../../messages';
-import { DeletedUser } from '../Notification';
 import NotificationWrapper from '../NotificationWrapper';
+import UserLink from '../UserLink';
 
 interface Props {
   notification: ICommentOnIdeaYouFollowNotificationData;
@@ -32,9 +31,6 @@ const CommentOnIdeaYouFollowNotification = memo<Props>((props) => {
 
   if (!isNilOrError(project)) {
     const inputTerm = getInputTerm(phases?.data);
-    const deletedUser =
-      isNilOrError(notification.attributes.initiating_user_first_name) ||
-      isNilOrError(notification.attributes.initiating_user_slug);
 
     return (
       <NotificationWrapper
@@ -56,17 +52,11 @@ const CommentOnIdeaYouFollowNotification = memo<Props>((props) => {
             petition: messages.userCommentedOnPetitionYouFollow,
           })}
           values={{
-            name: deletedUser ? (
-              <DeletedUser>
-                <FormattedMessage {...messages.deletedUser} />
-              </DeletedUser>
-            ) : (
-              <Link
-                to={`/profile/${notification.attributes.initiating_user_slug}`}
-                onClick={stopPropagation}
-              >
-                {notification.attributes.initiating_user_first_name}
-              </Link>
+            name: (
+              <UserLink
+                userName={notification.attributes.initiating_user_first_name}
+                userSlug={notification.attributes.initiating_user_slug}
+              />
             ),
           }}
         />
