+++ front/app/containers/MainHeader/Components/NotificationMenu/components/IdeaMarkedAsSpamNotification/index.tsx
@@ -5,12 +5,10 @@ import { IIdeaMarkedAsSpamNotificationData } from 'api/notifications/types';
 import T from 'components/T';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import Link from 'utils/cl-router/Link';
-import { isNilOrError, stopPropagation } from 'utils/helperUtils';
 
 import messages from '../../messages';
-import { DeletedUser } from '../Notification';
 import NotificationWrapper from '../NotificationWrapper';
+import UserLink from '../UserLink';
 
 interface Props {
   notification: IIdeaMarkedAsSpamNotificationData;
@@ -18,10 +16,6 @@ interface Props {
 
 const IdeaMarkedAsSpamNotification = memo<Props>((props) => {
   const { notification } = props;
-
-  const deletedUser =
-    isNilOrError(notification.attributes.initiating_user_first_name) ||
-    isNilOrError(notification.attributes.initiating_user_slug);
 
   return (
     <NotificationWrapper
@@ -33,17 +27,11 @@ const IdeaMarkedAsSpamNotification = memo<Props>((props) => {
       <FormattedMessage
         {...messages.userMarkedPostAsSpam}
         values={{
-          name: deletedUser ? (
-            <DeletedUser>
-              <FormattedMessage {...messages.deletedUser} />
-            </DeletedUser>
-          ) : (
-            <Link
-              to={`/profile/${notification.attributes.initiating_user_slug}`}
-              onClick={stopPropagation}
-            >
-              {notification.attributes.initiating_user_first_name}
-            </Link>
+          name: (
+            <UserLink
+              userName={notification.attributes.initiating_user_first_name}
+              userSlug={notification.attributes.initiating_user_slug}
+            />
           ),
           postTitle: <T value={notification.attributes.post_title_multiloc} />,
         }}
