+++ front/app/containers/MainHeader/Components/NotificationMenu/components/MentionInCommentNotification/index.tsx
@@ -3,12 +3,10 @@ import React, { memo } from 'react';
 import { IMentionInCommentNotificationData } from 'api/notifications/types';
 
 import { FormattedMessage } from 'utils/cl-intl';
-import Link from 'utils/cl-router/Link';
-import { isNilOrError, stopPropagation } from 'utils/helperUtils';
 
 import messages from '../../messages';
-import { DeletedUser } from '../Notification';
 import NotificationWrapper from '../NotificationWrapper';
+import UserLink from '../UserLink';
 
 interface Props {
   notification: IMentionInCommentNotificationData;
@@ -16,10 +14,6 @@ interface Props {
 
 const MentionInCommentNotification = memo<Props>((props) => {
   const { notification } = props;
-
-  const deletedUser =
-    isNilOrError(notification.attributes.initiating_user_first_name) ||
-    isNilOrError(notification.attributes.initiating_user_slug);
 
   return (
     <NotificationWrapper
@@ -31,17 +25,11 @@ const MentionInCommentNotification = memo<Props>((props) => {
       <FormattedMessage
         {...messages.mentionInComment}
         values={{
-          name: deletedUser ? (
-            <DeletedUser>
-              <FormattedMessage {...messages.deletedUser} />
-            </DeletedUser>
-          ) : (
-            <Link
-              to={`/profile/${notification.attributes.initiating_user_slug}`}
-              onClick={stopPropagation}
-            >
-              {notification.attributes.initiating_user_first_name}
-            </Link>
+          name: (
+            <UserLink
+              userName={notification.attributes.initiating_user_first_name}
+              userSlug={notification.attributes.initiating_user_slug}
+            />
           ),
         }}
       />
