+++ front/app/containers/MainHeader/Components/NotificationMenu/components/MentionInOfficialFeedbackNotification/index.tsx
@@ -9,7 +9,7 @@ import Link from 'utils/cl-router/Link';
 import { isNilOrError, stopPropagation } from 'utils/helperUtils';
 
 import messages from '../../messages';
-import { DeletedUser } from '../Notification';
+import { DeletedUser } from '../UserLink';
 import NotificationWrapper from '../NotificationWrapper';
 
 interface Props {
@@ -36,9 +36,7 @@ const MentionInCommentNotification = memo<Props>((props) => {
         {...messages.mentionInOfficialFeedback}
         values={{
           officialName: deletedUser ? (
-            <DeletedUser>
-              <FormattedMessage {...messages.deletedUser} />
-            </DeletedUser>
+            <DeletedUser />
           ) : (
             <Link
               to={`/profile/${notification.attributes.initiating_user_slug}`}
